clippy fix

+++ webschembly-compiler/src/lexer/located.rs
@@ -17,6 +17,6 @@ pub fn to_span(located: &LocatedStr) -> Span {
 pub fn to_pos(located: &LocatedStr) -> Pos {
     Pos::new(
         located.location_line() as usize,
-        located.get_utf8_column() as usize,
+        located.get_utf8_column(),
     )
 }
